• previously pressing tab would insert the common prefix of all items show. Now it looks only at items starting with the current one and to the end of the list.
After the (common) prefix has been inserted, the list is filtered causing all items above the current one to disappear, but the selection does not change. I did not add code to set the selection in the tab-handling method since this is a problem with all actions causing the list to be filtered (that the selection is not preserved), so likely the ‘filter’ method should handle this.


git-svn-id: http://svn.textmate.org/branches/WIP/Tools/Dialog2@10584 dfb7d73b-c2ec-0310-8fea-fb051d288c6d