• move image registration out of the popup command
This means it is now a two-step process to use images in popups (when those images are referring to disk files), esxample:

    "$DIALOG" register_images --images  "{ macro = '$(find_app com.macromates.textmate)/Contents/Resources/Bundle Item Icons/Macros.png'; }"
    "$DIALOG" popup --suggestions '( { image = macro; display = foo; }, { display = bar; } )'
	
But since images are registered with the given name, it means they can also be used in nibs, and so, this functionality is more general than just the popup stuff.

If an image is already registered under a requested name, the new image will not replace the existing one, so use unique/namespaced names for your images (although sematics may change).


git-svn-id: http://svn.textmate.org/branches/WIP/Tools/Dialog2@10878 dfb7d73b-c2ec-0310-8fea-fb051d288c6d